buildpack/deploy: fix issues with environment variables
I'm not sure why this issue happens now, but using exec /bin/bash -c
instead of plain exec fixes the issue. Before that, I was getting
"Error: '$PORT' is not a valid port number."

Now it works, and I can still load environment variables. I tried
removing the sed, but that didn't help with environment variables.
Probably there is a flag, let's see in the future :)